+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 1.0.0
 *In development*
 
 - Turn :meth:`Element.unserialize` into an instance method.
+- Rename :meth:`Element.set` to `Element.set_from_raw`.
 
 Version 0.2.0
 -------------